chore(button-timed): extract out use-paused
