+++ src/button-timed/button-timed.tsx
@@ -11,6 +11,7 @@ import {
   BaseButtonTimed as StyledBaseButtonTimed,
   TimerContainer as StyledTimerContainer,
 } from './styled-components';
+import { usePaused } from './use-paused';
 import { getOverrides, mergeOverrides } from '../helpers/overrides';
 
 function padTo2Digits(num) {
@@ -21,28 +22,6 @@ const formatTime = (totalSeconds) => {
   const seconds = totalSeconds % 60;
   return `${minutes}:${padTo2Digits(seconds)}`;
 };
-function usePrevious(value) {
-  const ref = React.useRef();
-  React.useEffect(() => {
-    ref.current = value;
-  });
-  return ref.current;
-}
-function useInterval(callback, delay) {
-  const intervalRef = React.useRef(null);
-  const savedCallback = React.useRef(callback);
-  React.useEffect(() => {
-    savedCallback.current = callback;
-  }, [callback]);
-  React.useEffect(() => {
-    const tick = () => savedCallback.current();
-    if (typeof delay === 'number') {
-      intervalRef.current = window.setInterval(tick, delay);
-      return () => window.clearInterval(intervalRef.current);
-    }
-  }, [delay]);
-  return intervalRef;
-}
 
 const ButtonTimed = ({
   initialTime,
@@ -53,18 +32,7 @@ const ButtonTimed = ({
   overrides = {},
   ...restProps
 }: ButtonTimedProps) => {
-  const [paused, setPaused] = React.useState<boolean>(false);
-  const prevTimeRemaining = usePrevious(timeRemaining);
-
-  useInterval(() => {
-    if (!paused && timeRemaining === prevTimeRemaining) {
-      setPaused(true);
-    }
-  }, 1000);
-
-  if (paused && timeRemaining !== prevTimeRemaining) {
-    setPaused(false);
-  }
+  const paused = usePaused(timeRemaining);
 
   const { TimerContainer: TimerContainerOverride, ...buttonOverrides } = overrides;
   const [TimerContainer, timerContainerProps] = getOverrides(
